move progress tracking to it's own module
